# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Femwell Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions:
    get:
      operationId: getUserSubscription
      tags: [subscriptions]
      summary: Get current user's subscription
      responses:
        "200":
          description: Subscription or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Subscribe to a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}/cancel:
    patch:
      operationId: cancelSubscription
      tags: [subscriptions]
      summary: Cancel a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  # ─── TELEHEALTH ──────────────────────────────────────────────────────────────
...
  /users/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Log in a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginInput"
      responses:
        "200":
          description: Auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── DASHBOARD (wow endpoints) ────────────────────────────────────────────────
...
  /tracker/entries:
    get:
      operationId: listCycleEntries
      tags: [tracker]
      summary: List all cycle entries for the current user
      responses:
        "200":
          description: Cycle entries list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CycleEntry"
    post:
      operationId: logCycleEntry
      tags: [tracker]
      summary: Log a new period/cycle entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCycleEntryBody"
      responses:
        "201":
          description: Cycle entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CycleEntry"
...
components:
  schemas:
...
    Subscription:
      type: object
      required: [id, userId, planId, planName, status, startDate, nextDelivery]
      properties:
        id:
          type: integer
        userId:
          type: integer
        planId:
          type: integer
        planName:
          type: string
        status:
          type: string
          enum: [active, paused, cancelled]
        startDate:
          type: string
        nextDelivery:
          type: string
        deliveryAddress:
          type: ["string", "null"]
    SubscriptionInput:
      type: object
      required: [planId, deliveryAddress, paymentMethod]
      properties:
        planId:
          type: integer
        deliveryAddress:
          type: string
        paymentMethod:
          type: string

    # ─── TELEHEALTH ──────────────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin]
        phone:
          type: ["string", "null"]
        university:
          type: ["string", "null"]
        deliveryAddress:
          type: ["string", "null"]
        createdAt:
          type: string
    UserRegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        phone:
          type: string
        university:
          type: string
    UserLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        university:
          type: string
        deliveryAddress:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ─── DASHBOARD ──────────────────────────────────────────────────────────────
...
    CycleEntry:
      type: object
      required: [id, userId, periodStart, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        periodStart:
          type: string
        periodEnd:
          type: ["string", "null"]
        cycleLength:
          type: ["integer", "null"]
        symptoms:
          type: array
          items:
            type: string
        flow:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    CreateCycleEntryBody:
      type: object
      required: [periodStart]
      properties:
        periodStart:
          type: string
        periodEnd:
          type: string
        cycleLength:
          type: integer
        symptoms:
          type: array
          items:
            type: string
        flow:
          type: string
          enum: [light, medium, heavy]
        notes:
          type: string
...