# lib/api-zod/src/generated/types/cycleEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Femwell Connect API
 * OpenAPI spec version: 0.1.0
 */

export interface CycleEntry {
  id: number;
  userId: number;
  periodStart: string;
  /** @nullable */
  periodEnd?: string | null;
  /** @nullable */
  cycleLength?: number | null;
  symptoms?: string[];
  /** @nullable */
  flow?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}