# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Femwell Connect API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  university?: string | null;
  /** @nullable */
  deliveryAddress?: string | null;
  createdAt: string;
}