# lib/api-zod/src/generated/types/userRegisterInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Femwell Connect API
 * OpenAPI spec version: 0.1.0
 */

export interface UserRegisterInput {
  email: string;
  password: string;
  name: string;
  phone?: string;
  university?: string;
}